Caches block hash for performance
The app would get increasingly slow as more blocks were mined. This is
because every time a block hash was output, it would require calculating
the hash for each block all the way back to the origin block. This is
good that this was happening - it means the algorithm works correctly.
(The hash for each block should be dependent on all preceding block
data.) By caching it though we can keep the program usable.