Store index of next block rather than the next block itself
If we store the block we're actually just storing a copy of it. When the
original is modified (once it gets "nextBlock" set on it), the record
that is linked to will not be updated.

The index serves as a reference. As long as the items in the list do not
change position (which they will not), then block can always be looked
up by the index, and it will be the actual up-to-date block, not an
out-of-date copy.